+++ views/index.py
@@ -326,7 +326,7 @@ def reserve(copy_id):
                 user_id=session['id'],
                 book_status=BookStatus.RESERVED,
                 reservation_begin=datetime.now(tz=pytz.utc),
-                reservation_end=datetime.now(tz=pytz.utc) + timedelta(hours=48)
+                reservation_end=datetime.now(tz=pytz.utc) + timedelta(minutes=2)
             )
             db.session.add(res)
             db.session.commit()
@@ -334,6 +334,24 @@ def reserve(copy_id):
         except IntegrityError:
             abort(500)
     return redirect(url_for('library.index'))
+
+
+@library.route('/check_reservation_status_db')
+def check_reservation_status_db():
+    reserved_list = db.session.query(RentalLog)\
+        .filter(RentalLog.book_status == BookStatus.RESERVED)\
+        .filter(RentalLog._reservation_end < datetime.utcnow())\
+        .all()
+    for reserved_row in reserved_list:
+        db.session.query(Copy)\
+            .filter(Copy.id == reserved_row.copy_id)\
+            .update({Copy.available_status: True})
+    db.session.query(RentalLog)\
+        .filter(RentalLog.book_status == BookStatus.RESERVED)\
+        .filter(RentalLog._reservation_end < datetime.utcnow())\
+        .update({RentalLog.book_status: BookStatus.RETURNED})
+    db.session.commit()
+    return "OK"
 
 
 @library.route('/remove_item/<int:item_id>', methods=['GET', 'POST'])