+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Add and Edit Copy feature
 - Flashing messages support (as Jinja macro)
+- Travis CI configuration
+- Reservation mechanism design and implement
 
 ## [0.0.1] - 2018-06-11
 ### Added
@@ -19,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fix bug - Eliminate bug with already registered user registration handling
 - Creating form for add books in wishlist
 - Creating like button in wishlist for users
+- Fix for available status
 
 ##[0.0.1] - 2018.06.11
 ### Added